+++ Source/Game.cpp
@@ -47,8 +47,19 @@ void Game::Start() {
 	ui->CreateFont("arial20px", TexMan::GetTex("arial20px"), "Textures/Interface/Fonts/arial20px.json");
 	ui->CreateFont("arial12px", TexMan::GetTex("arial12px"), "Textures/Interface/Fonts/arial12px.json");
 	//ui->CrateTempFontFromTTF("Fonts/arial.ttf", 12, "arial30px");
-
+	ui->settings.stopCheckAtFirst = true;
+	ui->settings.clickBoxStartAtDown = true;
 	renderer->FLatRenderCopySetUp();
+
+	ClickBox* cb = ui->CreateClickBoxF("first", 100, 100, 300, 300,nullptr,"arial40px","First Elem");
+	cb->SetRenderTextType(2);
+	cb->SetColor(30, 30, 30);
+	cb->SetHoverFilter(true, 255, 255, 255, 100);
+
+	cb = ui->CreateClickBoxF("second", 200, 200, 300, 300, nullptr, "arial40px", "Second Elem");
+	cb->SetRenderTextType(2);
+	cb->SetColor(255, 30, 30);
+	cb->SetHoverFilter(true, 255, 255, 255, 100);
 
 }
 
@@ -57,6 +68,13 @@ void Game::LogicUpdate() {
 }
 
 void Game::FrameUpdate() {
+	if (ui->ConsumeIfExist("first")) {
+		std::println("First clicked");
+	}
+	if (ui->ConsumeIfExist("second")) {
+		std::println("Second clicked");
+	}
+
 	Input();
 	ui->FrameUpdate();
 	Render();